Feat: Unify module loader into a single module loader (eszips), disect File system, graph and module loaders into their own crates. Clean up references to worker context, general refactors
