--- crates/base/src/js_worker/mod.rs
@@ -1,4 +1,0 @@
-pub mod emitter;
-pub mod module_loader;
-pub mod node_module_loader;
-pub mod standalone;
