+++ crates/base/src/lib.rs
@@ -1,13 +1,9 @@
 extern crate core;
 
-pub mod cert;
 pub mod commands;
 pub mod deno_runtime;
-pub mod errors_rt;
-pub mod js_worker;
 pub mod macros;
 pub mod rt_worker;
 pub mod server;
 pub mod snapshot;
-pub mod standalone;
 pub mod utils;
