--- crates/base/src/standalone/mod.rs
@@ -1,348 +1,0 @@
-// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-
-use deno_ast::MediaType;
-use deno_core::anyhow::Context;
-use deno_core::error::generic_error;
-use deno_core::error::type_error;
-use deno_core::error::AnyError;
-use deno_core::futures::FutureExt;
-use deno_core::ModuleLoader;
-use deno_core::ModuleSpecifier;
-use deno_core::ModuleType;
-use deno_core::ResolutionKind;
-use deno_fs;
-use deno_npm::NpmSystemInfo;
-use deno_semver::npm::NpmPackageReqReference;
-use deno_tls::rustls::RootCertStore;
-use deno_tls::RootCertStoreProvider;
-use import_map::ImportMap;
-use module_fetcher::args::package_json::PackageJsonDepsProvider;
-use module_fetcher::args::CacheSetting;
-use module_fetcher::cache::{Caches, DenoDirProvider, NodeAnalysisCache};
-use module_fetcher::file_fetcher::get_source_from_data_url;
-use module_fetcher::http_util::HttpClient;
-use module_fetcher::node::CliCjsCodeAnalyzer;
-use sb_node::analyze::NodeCodeTranslator;
-use std::pin::Pin;
-use std::rc::Rc;
-use std::sync::Arc;
-
-pub mod binary;
-pub mod file_system;
-pub mod virtual_fs;
-
-use crate::cert::{get_root_cert_store, CaData};
-use crate::deno_runtime::RuntimeProviders;
-use crate::js_worker::node_module_loader::{CjsResolutionStore, NpmModuleLoader};
-use crate::utils::graph_resolver::MappedSpecifierResolver;
-use sb_node::NodeResolver;
-use sb_npm::{
-    create_npm_fs_resolver, CliNpmRegistryApi, CliNpmResolver, NpmCache, NpmCacheDir, NpmResolution,
-};
-
-use self::binary::load_npm_vfs;
-use self::binary::Metadata;
-use self::file_system::DenoCompileFileSystem;
-
-pub const VFS_ESZIP_KEY: &str = "---SUPABASE-VFS-DATA-ESZIP---";
-
-pub struct SharedModuleLoaderState {
-    eszip: eszip::EszipV2,
-    mapped_specifier_resolver: MappedSpecifierResolver,
-    npm_module_loader: Arc<NpmModuleLoader>,
-}
-
-#[derive(Clone)]
-pub struct EmbeddedModuleLoader {
-    shared: Arc<SharedModuleLoaderState>,
-}
-
-impl ModuleLoader for EmbeddedModuleLoader {
-    fn resolve(
-        &self,
-        specifier: &str,
-        referrer: &str,
-        kind: ResolutionKind,
-    ) -> Result<ModuleSpecifier, AnyError> {
-        let referrer = if referrer == "." {
-            if kind != ResolutionKind::MainModule {
-                return Err(generic_error(format!(
-                    "Expected to resolve main module, got {:?} instead.",
-                    kind
-                )));
-            }
-            let current_dir = std::env::current_dir().unwrap();
-            deno_core::resolve_path(".", &current_dir)?
-        } else {
-            ModuleSpecifier::parse(referrer)
-                .map_err(|err| type_error(format!("Referrer uses invalid specifier: {}", err)))?
-        };
-
-        let permissions = sb_node::allow_all();
-        if let Some(result) = self.shared.npm_module_loader.resolve_if_in_npm_package(
-            specifier,
-            &referrer,
-            &*permissions,
-        ) {
-            return result;
-        }
-
-        let maybe_mapped = self
-            .shared
-            .mapped_specifier_resolver
-            .resolve(specifier, &referrer)?
-            .into_specifier();
-
-        // npm specifier
-        let specifier_text = maybe_mapped
-            .as_ref()
-            .map(|r| r.as_str())
-            .unwrap_or(specifier);
-        if let Ok(reference) = NpmPackageReqReference::from_str(specifier_text) {
-            return self
-                .shared
-                .npm_module_loader
-                .resolve_req_reference(&reference, &*permissions);
-        }
-
-        match maybe_mapped {
-            Some(resolved) => Ok(resolved),
-            None => {
-                deno_core::resolve_import(specifier, referrer.as_str()).map_err(|err| err.into())
-            }
-        }
-    }
-
-    fn load(
-        &self,
-        original_specifier: &ModuleSpecifier,
-        maybe_referrer: Option<&ModuleSpecifier>,
-        _is_dynamic: bool,
-    ) -> Pin<Box<deno_core::ModuleSourceFuture>> {
-        let is_data_uri = get_source_from_data_url(original_specifier).ok();
-        let permissions = sb_node::allow_all();
-        if let Some((source, _)) = is_data_uri {
-            return Box::pin(deno_core::futures::future::ready(Ok(
-                deno_core::ModuleSource::new(
-                    deno_core::ModuleType::JavaScript,
-                    source.into(),
-                    original_specifier,
-                ),
-            )));
-        }
-
-        if let Some(result) = self.shared.npm_module_loader.load_sync_if_in_npm_package(
-            original_specifier,
-            maybe_referrer,
-            &*permissions,
-        ) {
-            return match result {
-                Ok(code_source) => Box::pin(deno_core::futures::future::ready(Ok(
-                    deno_core::ModuleSource::new_with_redirect(
-                        match code_source.media_type {
-                            MediaType::Json => ModuleType::Json,
-                            _ => ModuleType::JavaScript,
-                        },
-                        code_source.code,
-                        original_specifier,
-                        &code_source.found_url,
-                    ),
-                ))),
-                Err(err) => Box::pin(deno_core::futures::future::ready(Err(err))),
-            };
-        }
-
-        let Some(module) = self.shared.eszip.get_module(original_specifier.as_str()) else {
-            return Box::pin(deno_core::futures::future::ready(Err(type_error(format!(
-                "Module not found: {}",
-                original_specifier
-            )))));
-        };
-        let original_specifier = original_specifier.clone();
-        let found_specifier =
-            ModuleSpecifier::parse(&module.specifier).expect("invalid url in eszip");
-
-        async move {
-            let code = module
-                .source()
-                .await
-                .ok_or_else(|| type_error(format!("Module not found: {}", original_specifier)))?;
-            let code =
-                arc_u8_to_arc_str(code).map_err(|_| type_error("Module source is not utf-8"))?;
-            Ok(deno_core::ModuleSource::new_with_redirect(
-                match module.kind {
-                    eszip::ModuleKind::JavaScript => ModuleType::JavaScript,
-                    eszip::ModuleKind::Json => ModuleType::Json,
-                    eszip::ModuleKind::Jsonc => {
-                        return Err(type_error("jsonc modules not supported"))
-                    }
-                    eszip::ModuleKind::OpaqueData => {
-                        unreachable!();
-                    }
-                },
-                code.into(),
-                &original_specifier,
-                &found_specifier,
-            ))
-        }
-        .boxed_local()
-    }
-}
-
-fn arc_u8_to_arc_str(arc_u8: Arc<[u8]>) -> Result<Arc<str>, std::str::Utf8Error> {
-    // Check that the string is valid UTF-8.
-    std::str::from_utf8(&arc_u8)?;
-    // SAFETY: the string is valid UTF-8, and the layout Arc<[u8]> is the same as
-    // Arc<str>. This is proven by the From<Arc<str>> impl for Arc<[u8]> from the
-    // standard library.
-    Ok(unsafe { std::mem::transmute(arc_u8) })
-}
-
-pub struct StandaloneModuleLoaderFactory {
-    shared: Arc<SharedModuleLoaderState>,
-}
-
-struct StandaloneRootCertStoreProvider {
-    ca_stores: Option<Vec<String>>,
-    ca_data: Option<CaData>,
-    cell: once_cell::sync::OnceCell<RootCertStore>,
-}
-
-impl RootCertStoreProvider for StandaloneRootCertStoreProvider {
-    fn get_or_try_init(&self) -> Result<&RootCertStore, AnyError> {
-        self.cell.get_or_try_init(|| {
-            get_root_cert_store(None, self.ca_stores.clone(), self.ca_data.clone())
-                .map_err(|err| err.into())
-        })
-    }
-}
-
-pub async fn create_module_loader_for_eszip(
-    mut eszip: eszip::EszipV2,
-    metadata: Metadata,
-    maybe_import_map: Option<ImportMap>,
-) -> Result<RuntimeProviders, AnyError> {
-    // let main_module = &metadata.entrypoint;
-    let current_exe_path = std::env::current_exe().unwrap();
-    let current_exe_name = current_exe_path.file_name().unwrap().to_string_lossy();
-    let deno_dir_provider = Arc::new(DenoDirProvider::new(None));
-    let root_cert_store_provider = Arc::new(StandaloneRootCertStoreProvider {
-        ca_stores: metadata.ca_stores,
-        ca_data: metadata.ca_data.map(CaData::Bytes),
-        cell: Default::default(),
-    });
-    let http_client = Arc::new(HttpClient::new(
-        Some(root_cert_store_provider.clone()),
-        metadata.unsafely_ignore_certificate_errors.clone(),
-    ));
-
-    // use a dummy npm registry url
-    let npm_registry_url = ModuleSpecifier::parse("https://localhost/").unwrap();
-    let root_path = std::env::temp_dir()
-        .join(format!("sb-compile-{}", current_exe_name))
-        .join("node_modules");
-    let npm_cache_dir = NpmCacheDir::new(root_path.clone());
-    let (fs, snapshot) = if let Some(snapshot) = eszip.take_npm_snapshot() {
-        // TODO: Support node_modules
-        let vfs_root_dir_path = npm_cache_dir.registry_folder(&npm_registry_url);
-
-        let vfs_data: Vec<u8> = eszip
-            .get_module(VFS_ESZIP_KEY)
-            .unwrap()
-            .take_source()
-            .await
-            .unwrap()
-            .to_vec();
-
-        let vfs = load_npm_vfs(vfs_root_dir_path, &vfs_data).context("Failed to load npm vfs.")?;
-
-        (
-            Arc::new(DenoCompileFileSystem::new(vfs)) as Arc<dyn deno_fs::FileSystem>,
-            Some(snapshot),
-        )
-    } else {
-        (
-            Arc::new(deno_fs::RealFs) as Arc<dyn deno_fs::FileSystem>,
-            None,
-        )
-    };
-
-    let npm_cache = Arc::new(NpmCache::new(
-        npm_cache_dir,
-        CacheSetting::Only,
-        fs.clone(),
-        http_client.clone(),
-    ));
-
-    let npm_api = Arc::new(CliNpmRegistryApi::new(
-        npm_registry_url.clone(),
-        npm_cache.clone(),
-        http_client.clone(),
-    ));
-
-    let npm_resolution = Arc::new(NpmResolution::from_serialized(
-        npm_api.clone(),
-        snapshot,
-        None,
-    ));
-
-    let npm_fs_resolver = create_npm_fs_resolver(
-        fs.clone(),
-        npm_cache,
-        npm_registry_url,
-        npm_resolution.clone(),
-        None,
-        NpmSystemInfo::default(),
-    );
-
-    let npm_resolver = Arc::new(CliNpmResolver::new(
-        fs.clone(),
-        npm_resolution.clone(),
-        npm_fs_resolver,
-        None,
-    ));
-
-    let node_resolver = Arc::new(NodeResolver::new(fs.clone(), npm_resolver.clone()));
-    let cjs_resolutions = Arc::new(CjsResolutionStore::default());
-    let cache_db = Caches::new(deno_dir_provider.clone());
-    let node_analysis_cache = NodeAnalysisCache::new(cache_db.node_analysis_db());
-    let cjs_esm_code_analyzer = CliCjsCodeAnalyzer::new(node_analysis_cache, fs.clone());
-    let node_code_translator = Arc::new(NodeCodeTranslator::new(
-        cjs_esm_code_analyzer,
-        fs.clone(),
-        node_resolver.clone(),
-        npm_resolver.clone(),
-    ));
-    let package_json_deps_provider = Arc::new(PackageJsonDepsProvider::new(
-        metadata
-            .package_json_deps
-            .map(|serialized| serialized.into_deps()),
-    ));
-    let maybe_import_map = maybe_import_map
-        .map(|import_map| Some(Arc::new(import_map)))
-        .unwrap_or_else(|| None);
-
-    let module_loader_factory = StandaloneModuleLoaderFactory {
-        shared: Arc::new(SharedModuleLoaderState {
-            eszip,
-            mapped_specifier_resolver: MappedSpecifierResolver::new(
-                maybe_import_map,
-                package_json_deps_provider.clone(),
-            ),
-            npm_module_loader: Arc::new(NpmModuleLoader::new(
-                cjs_resolutions,
-                node_code_translator,
-                fs.clone(),
-                node_resolver.clone(),
-            )),
-        }),
-    };
-
-    Ok(RuntimeProviders {
-        module_loader: Rc::new(EmbeddedModuleLoader {
-            shared: module_loader_factory.shared.clone(),
-        }),
-        npm_resolver,
-        fs,
-    })
-}
