+++ crates/base/tests/main_worker_tests.rs
@@ -1,6 +1,6 @@
 use base::rt_worker::worker_ctx::{create_user_worker_pool, create_worker};
 use hyper::{Body, Request, Response};
-use sb_worker_context::essentials::{
+use sb_workers::context::{
     MainWorkerRuntimeOpts, WorkerContextInitOpts, WorkerRequestMsg, WorkerRuntimeOpts,
 };
 use std::collections::HashMap;
