+++ crates/base/tests/worker_boot_tests.rs
@@ -1,9 +1,7 @@
 use std::collections::HashMap;
 
 use base::rt_worker::worker_ctx::create_worker;
-use sb_worker_context::essentials::{
-    UserWorkerRuntimeOpts, WorkerContextInitOpts, WorkerRuntimeOpts,
-};
+use sb_workers::context::{UserWorkerRuntimeOpts, WorkerContextInitOpts, WorkerRuntimeOpts};
 
 #[tokio::test]
 async fn test_worker_boot_invalid_imports() {
