+++ crates/module_fetcher/Cargo.toml
@@ -12,7 +12,7 @@ jsonc-parser = { version="0.21.0", features = ["serde"] }
 base32 = "=0.4.0"
 deno_npm.workspace = true
 deno_semver.workspace = true
-monch = "=0.4.3"
+monch.workspace = true
 walkdir = "=2.3.2"
 fs3.workspace = true
 regex.workspace = true
@@ -20,7 +20,7 @@ once_cell.workspace = true
 thiserror.workspace = true
 deno_lockfile.workspace = true
 indexmap.workspace = true
-async-trait = "0.1.73"
+async-trait.workspace = true
 base64 = { version = "=0.13.1" }
 cache_control = { version = "=0.2.0" }
 chrono = { version = "=0.4.22", default-features = false, features = ["clock"] }
