+++ crates/sb_core/lib.rs
@@ -1,3 +1,5 @@
+pub mod cert;
+pub mod errors_rt;
 pub mod http_start;
 pub mod net;
 pub mod permissions;
