--- crates/sb_eszip/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "sb_eszip"
-version = "0.1.0"
-authors = ["Supabase <team@supabase.com>"]
-edition = "2021"
-resolver = "2"
-license = "MIT"
-
-[lib]
-path = "lib.rs"
-
-[dependencies]
-anyhow.workspace = true
-deno_core.workspace = true
-eszip.workspace = true
-import_map = { version = "0.15.0" }
-log = { workspace = true }
-serde.workspace = true
-tokio.workspace = true
