--- crates/sb_eszip/lib.rs
@@ -1,1 +1,0 @@
-pub mod module_loader;
