+++ crates/sb_fs/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "sb_fs"
+version = "0.1.0"
+authors = ["Supabase <team@supabase.com>"]
+edition = "2021"
+resolver = "2"
+license = "MIT"
+
+[lib]
+path = "lib.rs"
+
+[dependencies]
+module_fetcher = { path = "../module_fetcher" }
+sb_node = { version = "0.1.0", path = "../node" }
+sb_npm = { version = "0.1.0", path = "../npm" }
+deno_semver.workspace = true
+anyhow.workspace = true
+deno_core.workspace = true
+eszip.workspace = true
+import_map = { version = "0.15.0" }
+log = { workspace = true }
+serde.workspace = true
+tokio.workspace = true
+deno_ast.workspace = true
+deno_fs.workspace = true
+deno_npm.workspace = true
+once_cell.workspace = true
+deno_io.workspace = true
+thiserror.workspace = true
+async-trait.workspace = true
