+++ crates/sb_module_loader/node/mod.rs
@@ -1,0 +1,1 @@
+pub mod node_module_loader;
