+++ crates/sb_module_loader/util.rs
@@ -1,0 +1,10 @@
+use std::sync::Arc;
+
+pub fn arc_u8_to_arc_str(arc_u8: Arc<[u8]>) -> Result<Arc<str>, std::str::Utf8Error> {
+    // Check that the string is valid UTF-8.
+    std::str::from_utf8(&arc_u8)?;
+    // SAFETY: the string is valid UTF-8, and the layout Arc<[u8]> is the same as
+    // Arc<str>. This is proven by the From<Arc<str>> impl for Arc<[u8]> from the
+    // standard library.
+    Ok(unsafe { std::mem::transmute(arc_u8) })
+}
