+++ crates/sb_os/Cargo.toml
@@ -14,5 +14,4 @@ path = "lib.rs"
 deno_core.workspace = true
 libc.workspace = true
 serde.workspace = true
-sb_worker_context = { version = "0.1.0", path = "../sb_worker_context" }
 sb_core = { version = "0.1.0", path = "../sb_core" }
