--- crates/sb_worker_context/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "sb_worker_context"
-version = "0.1.0"
-authors = ["Supabase <team@supabase.com>"]
-edition = "2021"
-resolver = "2"
-description = "We'll take care of this later"
-license = "MIT"
-
-[lib]
-path = "lib.rs"
-
-[dependencies]
-anyhow = { workspace = true }
-deno_core.workspace = true
-event_worker ={ version = "0.1.0", path = "../event_worker" }
-enum-as-inner = "0.6.0"
-hyper.workspace = true
-serde.workspace = true
-sb_eszip = { version = "0.1.0", path = "../sb_eszip" }
-tokio.workspace = true
-uuid.workspace = true
