--- crates/sb_worker_context/lib.rs
@@ -1,1 +1,0 @@
-pub mod essentials;
